nosql/python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
azure-core==1.30.1
    # via
    #   azure-cosmos
    #   azure-identity
    #   azure-storage-blob
azure-cosmos==4.6.0
    # via -r .\requirements.in
azure-identity==1.16.0
    # via -r .\requirements.in
azure-storage-blob==12.19.1
    # via -r .\requirements.in
black==24.4.2
    # via -r .\requirements.in
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
cryptography==42.0.5
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
distro==1.9.0
    # via openai
docopt==0.6.2
    # via -r .\requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   -r .\requirements.in
    #   openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
isodate==0.6.1
    # via azure-storage-blob
msal==1.28.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.1.0
    # via azure-identity
mypy-extensions==1.0.0
    # via black
openai==1.25.0
    # via -r .\requirements.in
packaging==24.0
    # via
    #   black
    #   msal-extensions
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.2.1
    # via black
pluggy==1.5.0
    # via pytest
portalocker==2.8.2
    # via msal-extensions
psutil==5.9.8
    # via -r .\requirements.in
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via openai
pydantic-core==2.18.2
    # via pydantic
pyjwt[crypto]==2.8.0
    # via
    #   msal
    #   pyjwt
pytest==8.2.0
    # via pytest-asyncio
pytest-asyncio==0.23.6
    # via -r .\requirements.in
python-dotenv==1.0.1
    # via -r .\requirements.in
pywin32==306
    # via portalocker
requests==2.31.0
    # via
    #   azure-core
    #   msal
six==1.16.0
    # via
    #   azure-core
    #   isodate
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tqdm==4.66.2
    # via openai
typing-extensions==4.11.0
    # via
    #   azure-core
    #   azure-cosmos
    #   azure-storage-blob
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
